requirements.txt
azure-keyvault-secrets
azure-identity
uvicorn[standard]
fastapi
pydantic
duo-client
